requirements.txt
flask
gunicorn
requests
pandas
scikit-learn
python-telegram-bot==13.15